[FIX] Players and synths can heal their blood back (#1540)
* FixedPoint2 isn't good enough for healing blood

* Synths can heal their blood

* float is smaller and should be enough resolution, shoutout Kenkaitos

* Extra space, pain

+++ Content.Shared/Body/Systems/SharedBloodstreamSystem.cs
@@ -502,7 +502,7 @@ public abstract partial class SharedBloodstreamSystem : EntitySystem
             return false;
 
         referenceFactor = Math.Clamp(referenceFactor, 0f, ent.Comp.MaxVolumeModifier);
-        var ratio = amount / ent.Comp.BloodReferenceSolution.Volume;
+        var ratio = (float)amount / (float)ent.Comp.BloodReferenceSolution.Volume; // Omu - FixedPoint2 only has 2 decimals of resolution
 
         foreach (var (referenceReagent, referenceQuantity) in ent.Comp.BloodReferenceSolution)
         {
